+++ plugins/newspack-plugin/src/wizards/audience/views/content-gates/types/index.d.ts
@@ -31,6 +31,7 @@ type AccessRule = {
 	options?: { value: string; label: string }[];
 	has_options: boolean;
 	empty_grants_access?: boolean;
+	requires_value?: boolean;
 	placeholder?: string;
 	value: GateAccessRuleValue;
 };